+++ src/embeddings.jl
@@ -1,29 +1,31 @@
 using StaticArrays
 using Base: @_inline_meta
-export reconstruct, DelayEmbedding, AbstractEmbedding, MTDelayEmbedding, embed
+export reconstruct, DelayEmbedding, MTDelayEmbedding, embed, τrange
 export WeightedDelayEmbedding
 
 #####################################################################################
-#                        Delay Embedding Reconstruction                             #
+# Univariate Delay Coordinates
 #####################################################################################
 """
     AbstractEmbedding
-Super-type of embedding methods. Use `subtypes(AbstractEmbedding)` for available
-methods.
+Super-type of embedding methods.
 """
-abstract type AbstractEmbedding <: Function end
+abstract type AbstractEmbedding end
 
 """
-    DelayEmbedding(γ, τ) -> `embedding`
+    DelayEmbedding(γ, τ) → `embedding`
 Return a delay coordinates embedding structure to be used as a functor,
 given a timeseries and some index. Calling
 ```julia
 embedding(s, n)
 ```
-will create the `n`-th reconstructed vector of the embedded space, which has `γ`
+will create the `n`-th delay vector of the embedded space, which has `γ`
 temporal neighbors with delay(s) `τ`. See [`reconstruct`](@ref) for more.
 
-**Be very careful when choosing `n`, because `@inbounds` is used internally.**
+**Be very careful when choosing `n`, because `@inbounds` is used internally.
+It must be that `n ≤ length(s) - maximum(τ)`.**
+
+Convience function [`τrange`](@ref) gives all valid `n` indices.
 """
 struct DelayEmbedding{γ} <: AbstractEmbedding
     delays::SVector{γ, Int}
@@ -52,12 +54,15 @@ end
 # Weighted version
 export WeightedDelayEmbedding
 """
-    WeightedDelayEmbedding(γ, τ, w) -> `embedding`
+    WeightedDelayEmbedding(γ, τ, w) → `embedding`
 Similar with [`DelayEmbedding`](@ref), but the entries of the
 embedded vector are further weighted with `w^γ`.
 See [`reconstruct`](@ref) for more.
 
-**Be very careful when choosing `n`, because `@inbounds` is used internally.**
+**Be very careful when choosing `n`, because `@inbounds` is used internally.
+It must be that `n ≤ length(s) - maximum(τ)`.**
+
+Convience function [`τrange`](@ref) gives all valid `n` indices.
 """
 struct WeightedDelayEmbedding{γ, T<:Real} <: AbstractEmbedding
     delays::SVector{γ, Int}
@@ -81,10 +86,10 @@ end
 """
     reconstruct(s, γ, τ [, w])
 Reconstruct `s` using the delay coordinates embedding with `γ` temporal neighbors
-and delay `τ` and return the result as a [`Dataset`](@ref).
+and delay `τ` and return the result as a [`Dataset`](@ref). Optionally use weight `w`.
 
 Use [`embed`](@ref) for the version that accepts the embedding dimension `D = γ+1`
-instead.
+instead. Here `τ ≥ 0`, use [`genembed`](@ref) for a generalized version.
 
 ## Description
 ### Single Timeseries
@@ -115,10 +120,7 @@ of the embedded vector are further weighted with ``w^\\gamma``, like so
 
 ### Multiple Timeseries
 To make a reconstruction out of a multiple timeseries (i.e. trajectory) the number
-of timeseries must be known by type, so `s` can be either:
-
-* `s::AbstractDataset{B}`
-* `s::SizedAray{A, B}`
+of timeseries must be known by type, so `s` must be a `Dataset`.
 
 If the trajectory is for example ``(x, y)`` and `τ` is integer, then the ``n``-th
 entry of the embedded space is
@@ -157,13 +159,19 @@ function reconstruct(s::AbstractVector{T}, γ, τ, w) where {T}
 end
 @inline function reconstruct(s::AbstractVector{T},
     de::Union{WeightedDelayEmbedding{γ}, DelayEmbedding{γ}}) where {T, γ}
-    L = length(s) - maximum(de.delays)
-    data = Vector{SVector{γ+1, T}}(undef, L)
-    @inbounds for i in 1:L
+    r = τrange(s, de)
+    data = Vector{SVector{γ+1, T}}(undef, length(r))
+    @inbounds for i in r
         data[i] = de(s, i)
     end
     return Dataset{γ+1, T}(data)
 end
+
+"""
+    τrange(s, de::AbstractEmbedding)
+Return the range `r` of valid indices `n` to create delay vectors out of `s` using `de`.
+"""
+τrange(s, de::AbstractEmbedding) = 1:(length(s) - maximum(de.delays))
 
 """
     embed(s, D, τ)
@@ -178,21 +186,24 @@ embed(s, D, τ) = reconstruct(s, D-1, τ)
 
 
 #####################################################################################
-#                              MultiDimensional R                                   #
+# Multiple timeseries
 #####################################################################################
 """
     MTDelayEmbedding(γ, τ, B) -> `embedding`
 Return a delay coordinates embedding structure to be used as a functor,
-given multiple timeseries (`B` in total), either as a [`Dataset`](@ref) or a
-`SizedArray`), and some index.
+that embeds multiple timeseries (`B` in total) given in the form of a [`Dataset`](@ref).
+
 Calling
 ```julia
 embedding(s, n)
 ```
-will create the `n`-th reconstructed vector of the embedded space, which has `γ`
-temporal neighbors with delay(s) `τ`. See [`reconstruct`](@ref) for more.
+where `s` is a `Dataset` will create the `n`-th delay vector of the embedded space,
+which has `γ` temporal neighbors with delay(s) `τ`. See [`reconstruct`](@ref) for more.
 
-**Be very careful when choosing `n`, because `@inbounds` is used internally.**
+**Be very careful when choosing `n`, because `@inbounds` is used internally.
+It must be that `n ≤ length(s) - maximum(τ)`.**
+
+Convience function [`τrange`](@ref) gives all valid `n` indices.
 """
 struct MTDelayEmbedding{γ, B, X} <: AbstractEmbedding
     delays::SMatrix{γ, B, Int, X} # X = γ*B = total dimension number
@@ -223,6 +234,7 @@ end
 @generated function (r::MTDelayEmbedding{γ, B, X})(
     s::Union{AbstractDataset{B, T}, SizedArray{Tuple{A, B}, T, 2, M}},
     i) where {γ, A, B, T, M, X}
+    typeof(s) <: SizedArray && @warn "Using SizedArrays is deprecated. Use Dataset instead."
     gensprev = [:(s[i, $d]) for d=1:B]
     gens = [:(s[i + r.delays[$k, $d], $d]) for k=1:γ for d=1:B]
     quote
@@ -230,6 +242,8 @@ end
         @inbounds return SVector{$(γ+1)*$B,T}($(gensprev...), $(gens...))
     end
 end
+
+τrange(s, de::MTDelayEmbedding) = 1:(size(s)[1] - maximum(de.delays))
 
 @inline function reconstruct(
     s::Union{AbstractDataset{B, T}, SizedArray{Tuple{A, B}, T, 2, M}},
@@ -245,13 +259,107 @@ end
     s::Union{AbstractDataset{B, T}, SizedArray{Tuple{A, B}, T, 2, M}},
     de::MTDelayEmbedding{γ, B, F}) where {A, B, T, M, γ, F}
 
-    L = size(s)[1] - maximum(de.delays)
+    r = τrange(s, de)
     X = (γ+1)*B
-    data = Vector{SVector{X, T}}(undef, L)
-    @inbounds for i in 1:L
+    data = Vector{SVector{X, T}}(undef, length(r))
+    @inbounds for i in r
         data[i] = de(s, i)
     end
     return Dataset{X, T}(data)
 end
 
 reconstruct(s::AbstractMatrix, args...) = reconstruct(Dataset(s), args...)
+
+#####################################################################################
+# Generalized embedding (arbitrary combination of timeseries and delays)
+#####################################################################################
+export GeneralizedEmbedding, genembed
+
+"""
+    GeneralizedEmbedding(τs, js) -> `embedding`
+Return a delay coordinates embedding structure to be used as a functor.
+Given a timeseries *or* trajectory (i.e. `Dataset`) `s` and calling
+```julia
+embedding(s, n)
+```
+will create the `n`-th delay vector of `s` in the embedded space using
+`generalized` embedding (see [`genembed`](@ref).
+
+`js` is ignored for timeseries input `s` (since all entries of `js` must be `1` in
+this case).
+
+**Be very careful when choosing `n`, because `@inbounds` is used internally.
+It must be that `minimum(τs) + 1 ≤ n ≤ length(s) - maximum(τs)`.
+In addition please ensure that all entries of `js` are valid dimensions of `s`.**
+
+Convience function [`τrange`](@ref) gives all valid `n` indices.
+"""
+struct GeneralizedEmbedding{D} <: AbstractEmbedding
+    τs::NTuple{D, Int}
+    js::NTuple{D, Int}
+end
+
+function Base.show(io::IO, g::GeneralizedEmbedding{D}) where {D}
+    print(io, "$D-dimensional generalized embedding\n")
+    print(io, "  τs: $(g.τs)\n")
+    print(io, "  js: $(g.js)")
+end
+
+# timeseries input
+@generated function (g::GeneralizedEmbedding{D})(s::AbstractArray{T}, i::Int) where {D, T}
+    gens = [:(s[i + g.τs[$k]]) for k=1:D]
+    quote
+        @_inline_meta
+        @inbounds return SVector{$D,T}($(gens...))
+    end
+end
+
+# dataset input
+@generated function (g::GeneralizedEmbedding{D})(s::Dataset{X, T}, i::Int) where {D, X, T}
+    gens = [:(s[i + g.τs[$k], g.js[$k]]) for k=1:D]
+    quote
+        @_inline_meta
+        @inbounds return SVector{$D,T}($(gens...))
+    end
+end
+
+τrange(s, ge::GeneralizedEmbedding) =
+max(1, (-minimum(ge.τs) + 1)):min(length(s), length(s) - maximum(ge.τs))
+
+
+"""
+    genembed(s, τs, js = ones(...)) → dataset
+Create a generalized embedding of `s` which can be a timeseries or arbitrary `Dataset`
+and return the result as a new `dataset`.
+
+The generalized embedding works as follows:
+- `τs::NTuple{D, Int}` denotes what delay times will be used for each of the entries
+  of the delay vector. It is strongly recommended that `τs[1] = 0`.
+  `τs` is allowed to have *negative entries* as well.
+- `js::NTuple{D, Int}` denotes which of the timeseries contained in `s`
+  will be used for the entries of the delay vector. `js` can contain duplicate indices.
+
+For example, imagine input trajectory ``s = [x, y, z]`` where ``x, y, z`` are timeseries
+(the columns of the `Dataset`).
+If `js = (1, 3, 2)` and `τs = (0, 2, -7)` the created delay vector at
+each step ``n`` will be
+```math
+(x(n), z(n+2), y(n-7))
+```
+
+`js` can be skipped, defaulting to index 1 (first timeseries) for all delay entries.
+
+See also [`reconstruct`](@ref). Internally uses [`GeneralizedEmbedding`](@ref).
+"""
+function genembed(s, τs::NTuple{D, Int}, js::NTuple{D, Int}) where {D}
+    ge::GeneralizedEmbedding{D} = GeneralizedEmbedding(τs, js)
+    r = τrange(s, ge)
+    T = eltype(s)
+    data = Vector{SVector{D, T}}(undef, length(r))
+    @inbounds for (i, n) in enumerate(r)
+        data[i] = ge(s, n)
+    end
+    return Dataset{D, T}(data)
+end
+
+genembed(s, τs::NTuple{D, Int}) where {D} = genembed(s, τs, NTuple{D, Int}(ones(D)))