+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# v1.4
+* New generalized embeddings: `genembed`, `GeneralizedEmbedding` that can do any combination possible from a dataset.
+* Using `SizedArrays` in multidimensional embeddings is deprecated and will be removed in later versions.
 # v1.3
 * Allow indexing datasets with boolean vectors.
 # v1.2.0